Remove circular dependency from `p2p.kademlia`
